\name{aic-methods}
\docType{methods}
\alias{aic-methods}
\alias{aic,ANY-method}
\title{Methods for Function aic }
\description{
 Used to compute the AIC value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{ Default method for e.g. \code{glm} and \code{lm}}
}}
\keyword{methods}
