\name{aicc-methods}
\docType{methods}
\alias{aicc-methods}
\alias{aicc,lm-method}
\title{  Methods for Function aicc}
\description{
 Used to compute the AICc value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "lm"}{ Method for \code{glm} and \code{lm} }
}}
\keyword{methods}

