\name{bic-methods}
\docType{methods}
\alias{bic-methods}
\alias{bic,lm-method}
\title{Methods for Function bic }
\description{
 Used to compute the QAICc value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "lm"}{ Method for \code{glm} and \code{lm}  }
}}
\keyword{methods}

