\name{getfit-methods}
\docType{methods}
\alias{getfit-methods}
\alias{getfit,ANY-method}
\title{ Methods for Function getfit}
\description{
 Used by \code{coef.glmulti} to access the contents of fitted model objects of different classes.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{ Default method, for e.g. \code{glm} objects. }

}}
\keyword{methods}
