\name{qaic-methods}
\docType{methods}
\alias{qaic-methods}
\alias{qaic,lm-method}
\alias{qaic,negbin-method}
\title{  Methods for Function qaic }
\description{
  Used to compute the QAIC value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "lm"}{ Method for \code{glm} and \code{lm}}
\item{object = "negbin"}{ Method for \code{glm.nb} }
}}
\keyword{methods}

