\name{qaicc-methods}
\docType{methods}
\alias{qaicc-methods}
\alias{qaicc,ANY-method}
\title{  Methods for Function qaicc }
\description{
 Used to compute the QAICc value for several classes of models.
 Specify the estimated c value to be used through the "glmulti-cvalue" option.

}
\section{Methods}{
\describe{

\item{object = "ANY"}{ Default method, for e.g. \code{glm} and \code{lm} }

}}
\keyword{methods}

