\name{weightable-methods}
\docType{methods}
\alias{weightable-methods}
\alias{weightable,glmulti-method}
\title{ Table of relative supports}
\description{
Prepares a table with model formulas, IC values and IC relative supports
}
\section{Methods}{
\describe{
	\item{object="glmulti"}{Default call}
}}
\keyword{methods}

