\name{weightable}
\alias{weightable}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Table of relative supports}
\description{
Prepares a table with model formulas, IC values and IC relative supports
}
\usage{
weightable(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ A \code{glmulti} object}
  \item{\dots}{For further improvements}
 }

\value{
 A data frame with the list of formulas, IC values and IC weights
}

\author{Jarret Byrnes and Vincent Calcagno, McGill University}

\seealso{ 

\code{\link{glmulti}}

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{models }
\keyword{regression }

