\name{write-methods}
\docType{methods}
\alias{write-methods}
\alias{write,glmulti-method}
\title{  Methods for Function write }
\description{
Methods for function \code{write} 
}
\section{Methods}{
\describe{

\item{x = "glmulti"}{ Method to write \code{glmulti} objects. }
}}
\keyword{methods}
