\name{aic-methods}
\docType{methods}
\alias{aic-methods}
\alias{aic,lm-method}
\alias{aic,negbin-method}
\title{Methods for Function aic }
\description{
 Used to compute the AIC value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "lm"}{ Method for \code{glm} and \code{lm}}
}}
\keyword{methods}
