\name{aicc-methods}
\docType{methods}
\alias{aicc-methods}
\alias{aicc,ANY-method}
\title{  Methods for Function aicc}
\description{
 Used to compute the AICc value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{ Default method for e.g. \code{glm} and \code{lm} }
}}
\keyword{methods}

