\name{bic-methods}
\docType{methods}
\alias{bic-methods}
\alias{bic,lm-method}
\alias{bic,negbin-method}
\title{Methods for Function bic }
\description{
 Used to compute the BIC value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "lm"}{ Method for \code{glm} and \code{lm}  }
\item{object = "negbin"}{ Method for \code{glm.nb} }
}}
\keyword{methods}

