\name{bic-methods}
\docType{methods}
\alias{bic-methods}
\alias{bic,ANY-method}
\title{Methods for Function bic }
\description{
 Used to compute the BIC value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{ Default method for e.g. \code{glm} and \code{lm}  }
}}
\keyword{methods}

