\name{getfit-methods}
\docType{methods}
\alias{getfit-methods}
\alias{getfit,glm-method}
\alias{getfit,lm-method}
\alias{getfit,negbin-method}
\title{ Methods for Function getfit}
\description{
 Used by \code{coef.glmulti} to access the contents of fitted model objects of different classes.
}
\section{Methods}{
\describe{

\item{object = "glm"}{ Method for \code{glm} objects. }

\item{object = "lm"}{ Method for \code{lm} objects.}

\item{object = "negbin"}{ Method for \code{glm.nb} objects.}
}}
\keyword{methods}
