\name{qaic-methods}
\docType{methods}
\alias{qaic-methods}
\alias{qaic,ANY-method}
\title{  Methods for Function qaic }
\description{
  Used to compute the QAIC value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{ Default method for e.g. \code{glm} and \code{lm}}
}}
\keyword{methods}

