\name{qaicc-methods}
\docType{methods}
\alias{qaicc-methods}
\alias{qaicc,lm-method}
\alias{qaicc,negbin-method}
\title{  Methods for Function qaicc }
\description{
 Used to compute the QAICc value for several classes of models.
}
\section{Methods}{
\describe{

\item{object = "lm"}{ Method for \code{glm} and \code{lm} }
\item{object = "negbin"}{ Method for \code{glm.nb} }
}}
\keyword{methods}

